# SakhiSafe-Backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/debug-permissions": {
      "get": {
        "operationId": "AuthController_debugPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/modules": {
      "get": {
        "operationId": "ModulesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "ModulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/modules/{id}": {
      "get": {
        "operationId": "ModulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      },
      "patch": {
        "operationId": "ModulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      },
      "delete": {
        "operationId": "ModulesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/modules/{id}/toggle": {
      "patch": {
        "operationId": "ModulesController_toggle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/roles/{roleId}/permissions": {
      "get": {
        "operationId": "PermissionsController_findByRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permissions"
        ]
      },
      "put": {
        "operationId": "PermissionsController_updateRolePermissions",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permissions"
        ]
      }
    },
    "/api/v1/organizations": {
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/care-seekers": {
      "get": {
        "operationId": "CareSeekersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "care-seekers"
        ]
      },
      "post": {
        "operationId": "CareSeekersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCareSeekerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "care-seekers"
        ]
      }
    },
    "/api/v1/care-seekers/{id}": {
      "get": {
        "operationId": "CareSeekersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "care-seekers"
        ]
      },
      "patch": {
        "operationId": "CareSeekersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCareSeekerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "care-seekers"
        ]
      }
    },
    "/api/v1/cases": {
      "get": {
        "operationId": "CasesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cases"
        ]
      },
      "post": {
        "operationId": "CasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{id}": {
      "get": {
        "operationId": "CasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cases"
        ]
      },
      "patch": {
        "operationId": "CasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cases"
        ]
      }
    },
...
  },
  "info": {
    "title": "SakhiSafe API",
    "description": "SakhiSafe backend API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@sakhisafe.local"
          },
          "password": {
            "type": "string",
            "example": "Admin@12345"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "DASHBOARD",
              "USERS",
              "ROLES",
              "MODULES",
              "ORGANIZATIONS",
              "CARE_SEEKERS",
              "CASES",
              "MESSAGES",
              "INCIDENTS",
              "SAFETY_LOGS",
              "EVIDENCE",
              "RAG_DOCUMENTS",
              "AUDIT_LOGS",
              "SYSTEM_SETTINGS"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "name"
        ]
      },
      "UpdateModuleDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "DASHBOARD",
              "USERS",
              "ROLES",
              "MODULES",
              "ORGANIZATIONS",
              "CARE_SEEKERS",
              "CASES",
              "MESSAGES",
              "INCIDENTS",
              "SAFETY_LOGS",
              "EVIDENCE",
              "RAG_DOCUMENTS",
              "AUDIT_LOGS",
              "SYSTEM_SETTINGS"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          }
        }
      },
      "RolePermissionItemDto": {
        "type": "object",
        "properties": {
          "moduleKey": {
            "type": "string",
            "enum": [
              "DASHBOARD",
              "USERS",
              "ROLES",
              "MODULES",
              "ORGANIZATIONS",
              "CARE_SEEKERS",
              "CASES",
              "MESSAGES",
              "INCIDENTS",
              "SAFETY_LOGS",
              "EVIDENCE",
              "RAG_DOCUMENTS",
              "AUDIT_LOGS",
              "SYSTEM_SETTINGS"
            ]
          },
          "canView": {
            "type": "boolean"
          },
          "canCreate": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean"
          }
        },
        "required": [
          "moduleKey",
          "canView",
          "canCreate",
          "canUpdate",
          "canDelete"
        ]
      },
      "UpdateRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePermissionItemDto"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "NGO",
              "GOVERNMENT",
              "ERT",
              "HEALTHCARE",
              "LEGAL",
              "POLICE",
              "SHELTER",
              "SYSTEM"
            ]
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "NGO",
              "GOVERNMENT",
              "ERT",
              "HEALTHCARE",
              "LEGAL",
              "POLICE",
              "SHELTER",
              "SYSTEM"
            ]
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "CreateCareSeekerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "safetyNotes": {
            "type": "string"
          }
        },
        "required": [
          "fullName"
        ]
      },
      "UpdateCareSeekerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "safetyNotes": {
            "type": "string"
          }
        }
      },
      "CreateCaseDto": {
        "type": "object",
        "properties": {
          "careSeekerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "assignedToId": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "incidentDescription": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "ESCALATED",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          }
        },
        "required": [
          "careSeekerId",
          "title"
        ]
      },
      "UpdateCaseDto": {
        "type": "object",
        "properties": {
          "careSeekerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "assignedToId": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "incidentDescription": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "ESCALATED",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          }
        }
      }
    }
  }
}